+++ tests/client/e2eTests/tests/users/page.js
@@ -4,10 +4,14 @@ var chai = require('chai');
 var chaiAsPromised = require('chai-as-promised');
 var e2e = require('@openveo/test').e2e;
 var UserPage = process.require('tests/client/e2eTests/pages/UserPage.js');
+var RolePage = process.require('tests/client/e2eTests/pages/RolePage.js');
 var UserModel = process.require('app/server/models/UserModel.js');
+var RoleModel = process.require('app/server/models/RoleModel.js');
 var UserProvider = process.require('app/server/providers/UserProvider.js');
+var RoleProvider = process.require('app/server/providers/RoleProvider.js');
 var storage = process.require('app/server/storage.js');
 var UserHelper = process.require('tests/client/e2eTests/helpers/UserHelper.js');
+var RoleHelper = process.require('tests/client/e2eTests/helpers/RoleHelper.js');
 var datas = process.require('tests/client/e2eTests/resources/data.json');
 var TableAssert = e2e.asserts.TableAssert;
 
@@ -16,51 +20,29 @@ var assert = chai.assert;
 chai.use(chaiAsPromised);
 
 describe('User page', function() {
-  var page, tableAssert, defaultUsers, userHelper;
-
-  /**
-   * Verifies roles of a user.
-   *
-   * @param {String} name The user name
-   * @param {Array} expectedUserRoles Expected list of roles expected for the user
-   * @return {Promise} Promise resolving when check is done
-   */
-  function checkRoles(name, expectedUserRoles) {
-    return page.getUserRoles(name).then(function(roles) {
-      var j = 0;
-
-      for (var i = 0; i < roles.length; i++) {
-        var role = roles[i];
-        if (expectedUserRoles.indexOf(role) >= 0)
-          j++;
-      }
-
-      assert.equal(j, expectedUserRoles.length, 'Roles of user ' + name + ' are incorrect');
-    });
-  }
-
-  /**
-   * Verifies email of a user.
-   *
-   * @param {String} name The user name
-   * @param {String} expectedEmail Expected user email
-   * @return {Promise} Promise resolving when check is done
-   */
-  function checkEmail(name, expectedEmail) {
-    return page.getLineFieldText(name, 'email').then(function(email) {
-      assert.equal(email, expectedEmail, 'User email is incorrect');
-    });
-  }
+  var page;
+  var tableAssert;
+  var defaultUsers;
+  var defaultRoles;
+  var userHelper;
+  var roleHelper;
+  var rolePage;
 
   // Prepare page
   before(function() {
     var userModel = new UserModel(new UserProvider(storage.getDatabase()));
+    var roleModel = new RoleModel(new RoleProvider(storage.getDatabase()));
     userHelper = new UserHelper(userModel);
+    roleHelper = new RoleHelper(roleModel);
     page = new UserPage(userModel);
+    rolePage = new RolePage(roleModel);
     tableAssert = new TableAssert(page, userHelper);
     page.logAsAdmin();
     userHelper.getEntities().then(function(users) {
       defaultUsers = users;
+    });
+    roleHelper.getEntities().then(function(roles) {
+      defaultRoles = roles;
     });
     page.load();
   });
@@ -71,7 +53,9 @@ describe('User page', function() {
 
   // Remove all extra users after each test and reload the page
   afterEach(function() {
+    process.protractorConf.startOpenVeo();
     userHelper.removeAllEntities(defaultUsers);
+    roleHelper.removeAllEntities(defaultRoles);
     page.refresh();
   });
 
@@ -83,124 +67,361 @@ describe('User page', function() {
     assert.eventually.ok(page.pageDescriptionElement.isPresent());
   });
 
+  it('should propose all available roles', function() {
+    roleHelper.getRoles().then(function(roles) {
+      assert.eventually.sameMembers(page.getAvailableRoles(), roles);
+    });
+  });
+
+  it('should propose new roles added from /be/roles-list page', function() {
+    var name = 'Test adding new role';
+    rolePage.translations = page.translations;
+
+    // Load roles page
+    page.clickMenu(page.translations.CORE.MENU.ROLES);
+
+    // Add a role
+    rolePage.addLine(name);
+
+    // Go back to users page
+    rolePage.clickMenu(rolePage.translations.CORE.MENU.USERS);
+
+    assert.eventually.include(page.getAvailableRoles(), name);
+  });
+
   it('should be able to add / remove a user', function() {
     var name = 'test add / remove user';
     var email = 'test-add-remove@veo-labs.com';
     var password = 'test-add-remove-password';
-    var roles = [datas.roles.coreAdmin.name, datas.roles.coreGuest.name];
 
-    // Add new user
-    page.addLine(name, {
-      email: email,
-      password: password,
-      passwordValidate: password,
-      roles: roles
+    roleHelper.getRoles().then(function(roles) {
+
+      // Add new user
+      page.addLine(name, {
+        email: email,
+        password: password,
+        passwordValidate: password,
+        roles: roles
+      });
+
+      // Verify user information
+      assert.eventually.equal(page.getLineFieldText(name, 'name'), name);
+      assert.eventually.sameMembers(page.getUserRoles(name), roles);
+      assert.eventually.equal(page.getLineFieldText(name, 'email'), email);
+
+      // Test connection with the new user
+      page.logAs({
+        name: name,
+        email: email,
+        password: password
+      });
+
+      // Log back to admin
+      page.logAsAdmin();
+      page.load();
+
+      // Remove user
+      page.removeLine(name);
+
+      assert.isRejected(page.getLine(name));
     });
 
-    // Verify user information
-    assert.isFulfilled(page.getLine(name));
-    checkRoles(name, roles);
-    checkEmail(name, email);
-
-    // Test connection with the new user
-    page.logAs({
-      name: name,
-      email: email,
-      password: password
-    });
-
-    // Log back to admin
-    page.logAsAdmin();
-    page.load();
-
-    // Remove user
-    page.removeLine(name);
   });
 
-  it('should not be able to add a user without name, email, password or password confirmation', function() {
-    var name = 'test add user without all info';
-    var email = 'test-add-without-all-info@veo-labs.com';
-    var password = 'test-add-without-all-info';
+  it('should not be able to add a user without name', function() {
+    var email = 'test-add-without-name@veo-labs.com';
+    var password = 'test-add-without-name';
 
-    assert.isRejected(page.addLine(name, {}));
-    assert.isRejected(page.addLine(name, {email: email}));
-    assert.isRejected(page.addLine(name, {email: 'invalidEmail'}));
+    assert.isRejected(page.addLine(null, {email: email, password: password, passwordValidate: password}));
+    assert.eventually.notOk(page.addButtonElement.isEnabled());
+  });
+
+  it('should not be able to add a user without email', function() {
+    var name = 'test add user without email';
+    var password = 'test-add-without-email';
+
+    assert.isRejected(page.addLine(name, {password: password, passwordValidate: password}));
+    assert.eventually.notOk(page.addButtonElement.isEnabled());
+  });
+
+  it('should not be able to add a user without password', function() {
+    var name = 'test add user without password';
+    var email = 'test-add-without-password@veo-labs.com';
+    var password = 'test-add-without-password';
+
+    assert.isRejected(page.addLine(name, {email: email, passwordValidate: password}));
+    assert.eventually.notOk(page.addButtonElement.isEnabled());
+  });
+
+  it('should not be able to add a user without password confirmation', function() {
+    var name = 'test add user without password confirmation';
+    var email = 'test-add-without-password-confirmation@veo-labs.com';
+    var password = 'test-add-without-password-confirmation';
+
+    assert.isRejected(page.addLine(name, {email: email, password: password}));
+    assert.eventually.notOk(page.addButtonElement.isEnabled());
+  });
+
+  it('should display an error notification if password confirmation does not match password', function() {
+    var name = 'test add user with wrong password confirmation';
+    var email = 'test-add-with-wrong-password-confirmation@veo-labs.com';
+    var password = 'test-add-with-wrong-password-confirmation';
+
     page.addLine(name, {email: email, password: password, passwordValidate: 'wrong'});
     page.getAlertMessages().then(function(messages) {
       assert.equal(messages.length, 1);
+      assert.ok(messages[0].indexOf(page.translations.CORE.ERROR.SERVER) === 0);
     });
     page.closeAlerts();
   });
 
-  it('should not display buttons to change the number of items per page if users lower than 6', function() {
-    page.getTotalLines().then(function(totalLines) {
-      if (totalLines < 6)
-        assert.eventually.equal(page.itemsPerPageLinkElements.count(), 0);
-    });
+  it('should display an error if adding a user failed on server side', function() {
+    var name = 'test add error';
+    var email = 'test-add-error@veo-labs.com';
+    var password = 'test-add-error';
+
+    process.protractorConf.stopOpenVeo();
+    page.addLine(name, {email: email, password: password, passwordValidate: password});
+
+    assert.eventually.sameMembers(page.getAlertMessages(), [page.translations.CORE.ERROR.SERVER]);
+    page.closeAlerts();
+
+    process.protractorConf.startOpenVeo();
+    assert.isRejected(page.getLine(name));
+  });
+
+  it('should display an error if removing a user failed on server side', function() {
+    var name = 'test remove error';
+    var email = 'test-remove-error@veo-labs.com';
+    var password = 'test-remove-error';
+
+    page.addLine(name, {email: email, password: password, passwordValidate: password});
+
+    // Search for the line before stopping the server
+    page.search({query: name});
+
+    process.protractorConf.stopOpenVeo();
+    page.removeLine(name);
+
+    assert.eventually.sameMembers(page.getAlertMessages(), [page.translations.CORE.ERROR.SERVER]);
+    page.closeAlerts();
+
+    assert.isFulfilled(page.getLine(name));
   });
 
   it('should be able to edit a user', function() {
     var name = 'test edition';
     var newName = 'test edition renamed';
     var newEmail = 'test-edition-renamed@veo-labs.com';
-    var newRoles = [datas.roles.coreGuest.name];
+    var password = 'test edition';
 
-    // Create user
-    page.addLine(name, {
-      email: 'test-edition@veo-labs.com',
-      password: 'test-edition',
-      passwordValidate: 'test-edition',
-      roles: [datas.roles.coreAdmin.name, datas.roles.coreGuest.name]
+    roleHelper.getRoles().then(function(newRoles) {
+
+      // Create user
+      page.addLine(name, {
+        email: 'test-edition@veo-labs.com',
+        password: password,
+        passwordValidate: password
+      });
+
+      // Edit user
+      page.editUser(name, {name: newName, email: newEmail, roles: newRoles});
+
+      // Verify user information
+      assert.eventually.equal(page.getLineFieldText(newName, 'name'), newName);
+      assert.eventually.sameMembers(page.getUserRoles(newName), newRoles);
+      assert.eventually.equal(page.getLineFieldText(newName, 'email'), newEmail);
+
+      // Test connection with new email
+      page.logAs({
+        name: newName,
+        email: newEmail,
+        password: password
+      });
+
+      // Log back to admin
+      page.logAsAdmin();
+      page.load();
+
     });
-
-    // Edit user
-    page.editUser(name, {name: newName, email: newEmail, roles: newRoles});
-
-    // Verify user information
-    assert.isFulfilled(page.getLine(newName));
-    checkRoles(newName, newRoles);
-    checkEmail(newName, newEmail);
   });
 
-  it('should not be able to edit a user without name or email', function() {
-    var name = 'test edit user without all info';
-    var email = 'test-edit-without-all-info@veo-labs.com';
-    var password = 'test-edit-without-all-info';
-    var roles = [datas.roles.coreAdmin.name, datas.roles.coreGuest.name];
+  it('should be able to cancel when editing a user', function() {
+    var name = 'test cancelling edition';
+    var newName = 'test cancelling edition edited';
+    var email = 'test-cancelling-edition@veo-labs.com';
+    var newEmail = 'test-cancelling-edition-edited@veo-labs.com';
 
-    // Add new user
-    page.addLine(name, {
-      email: email,
-      password: password,
-      passwordValidate: password,
-      roles: roles
+    roleHelper.getRoles().then(function(newRoles) {
+
+      // Create user
+      page.addLine(name, {
+        email: email,
+        password: 'test-cancelling-edition',
+        passwordValidate: 'test-cancelling-edition'
+      });
+
+      // Edit user
+      page.editUser(name, {name: newName, email: newEmail, roles: newRoles}, true);
+      assert.eventually.ok(page.isOpenedLine(name));
+
+      assert.eventually.equal(page.getLineFieldText(name, 'name'), name);
+      assert.eventually.equal(page.getLineFieldText(name, 'email'), email);
+      assert.eventually.equal(page.getLineFieldText(name, 'roles'), page.translations.CORE.UI.EMPTY);
     });
 
-    assert.isRejected(page.editUser(name, {name: '', email: ''}));
+  });
+
+  it('should indicate "Not set" if no roles for the user', function() {
+    var name = 'test add without roles';
+    var email = 'test-add-without-roles@veo-labs.com';
+    var password = 'test-add-without-roles';
+
+    page.addLine(name, {email: email, password: password, passwordValidate: password});
+    assert.eventually.equal(page.getLineFieldText(name, 'roles'), page.translations.CORE.UI.EMPTY);
+    page.removeLine(name);
+  });
+
+  it('should not be able to update a user without name', function() {
+    var name = 'test edit user without name';
+    var email = 'test-edit-without-name@veo-labs.com';
+    var password = 'test-edit-without-name';
+
+    page.addLine(name, {email: email, password: password, passwordValidate: password});
+    assert.isRejected(page.editUser(name, {name: ''}));
+  });
+
+  it('should not be able to update a user without email', function() {
+    var name = 'test edit user without email';
+    var email = 'test-edit-without-email@veo-labs.com';
+    var password = 'test-edit-without-email';
+
+    // Add new user
+    page.addLine(name, {email: email, password: password, passwordValidate: password});
+    assert.isRejected(page.editUser(name, {email: ''}));
+  });
+
+  it('should not be able to remove the super administrator', function() {
+    assert.eventually.sameMembers(page.getLineActions(e2e.users.testSuperAdmin.name), [
+      page.translations.CORE.UI.NO_ACTION
+    ]);
+  });
+
+  it('should propose actions available for all selected lines', function() {
+    var baseName = 'test common actions';
+
+    userHelper.addEntities([
+      {
+        name: baseName + ' not locked',
+        email: 'test-common-action-not-locked@veo-labs.com',
+        password: 'test-common-action-not-locked',
+        passwordValidate: 'test-common-action-not-locked'
+      },
+      {
+        name: baseName + ' locked',
+        email: 'test-common-action-locked@veo-labs.com',
+        password: 'test-common-action-locked',
+        passwordValidate: 'test-common-action-locked',
+        locked: true
+      }
+    ]);
+
+    page.search({query: baseName});
+
+    page.selectAllLines();
+    assert.eventually.sameMembers(page.getGlobalActions(), [
+      page.translations.CORE.UI.NO_COMMON_ACTION
+    ]);
+  });
+
+  it('should not be able to remove a locked user', function() {
+    page.selectLine(datas.users.coreLocked.name);
+    assert.eventually.sameMembers(page.getGlobalActions(), [
+      page.translations.CORE.UI.NO_COMMON_ACTION
+    ]);
+
+    assert.eventually.sameMembers(page.getLineActions(datas.users.coreLocked.name), [
+      page.translations.CORE.UI.NO_ACTION
+    ]);
+  });
+
+  it('should refuse a removed user', function() {
+    var name = 'test removed user';
+    var email = 'test-removed-user@veo-labs.com';
+    var password = 'test-removed-user';
+
+    page.addLine(name, {email: email, password: password, passwordValidate: password});
+    page.removeLine(name);
+
+    assert.isRejected(page.logAs({
+      name: name,
+      email: email,
+      password: password
+    }));
+
+    page.logAsAdmin();
+    page.load();
+  });
+
+  it('should logout a connected user when removed', function() {
+    var name = 'test logged out user';
+    var email = 'test-logged-out-user@veo-labs.com';
+    var password = 'test-logged-out-user';
+
+    roleHelper.getRoles().then(function(roles) {
+
+      page.addLine(name, {
+        name: name,
+        email: email,
+        password: password,
+        passwordValidate: 'test-logged-out-user',
+        roles: roles
+      });
+
+      page.logAs({
+        name: name,
+        email: email,
+        password: password
+      });
+      page.load();
+
+      page.removeLine(name);
+
+      assert.eventually.equal(browser.getCurrentUrl(), process.protractorConf.baseUrl + 'be/login');
+
+      page.logAsAdmin();
+      page.load();
+    });
   });
 
   it('should be able to cancel when removing a user', function() {
-    return tableAssert.checkCancelRemove(page.translations.CORE.USERS.QUERY_FILTER);
+    tableAssert.checkCancelRemove(page.translations.CORE.USERS.QUERY_FILTER);
   });
 
   it('should be able to sort by name', function() {
-    return tableAssert.checkSort(page.translations.CORE.USERS.NAME_COLUMN);
+    tableAssert.checkSort(page.translations.CORE.USERS.NAME_COLUMN);
   });
 
   it('should have buttons to change the number of items per page', function() {
-    return tableAssert.checkItemsPerPage();
+    tableAssert.checkItemsPerPage();
   });
 
   it('should be able to remove several lines simultaneously', function() {
-    return tableAssert.checkMassiveRemove();
+    tableAssert.checkMassiveRemove();
   });
 
   it('should be paginated', function() {
-    return tableAssert.checkPagination();
+    tableAssert.checkPagination();
   });
 
   it('should be able to select lines', function() {
-    return tableAssert.checkLinesSelection(page.translations.CORE.USERS.NAME_COLUMN);
+    tableAssert.checkLinesSelection(page.translations.CORE.USERS.NAME_COLUMN);
+  });
+
+  it('should have actions to remove users', function() {
+    tableAssert.checkActions([
+      page.translations.CORE.UI.REMOVE
+    ]);
   });
 
   describe('search', function() {